implement software interrupt in arm

+++ include/sim/system.h
@@ -9,7 +9,7 @@
 #ifndef __ASSEMBLY__
 
 
-static inline void generate_software_interrupt(volatile uint32_t id)
+static inline void generate_software_interrupt(volatile unsigned long id)
 {
   /* TODO: implement */
 }
